move collectFields into executor
to enable more memoization
