+++ src/execution/executor.ts
@@ -14,9 +14,14 @@ import type {
   FieldNode,
   FragmentDefinitionNode,
   OperationTypeNode,
+  SelectionSetNode,
+  FragmentSpreadNode,
+  InlineFragmentNode,
 } from 'graphql';
 
 import {
+  GraphQLIncludeDirective,
+  GraphQLSkipDirective,
   GraphQLError,
   Kind,
   SchemaMetaFieldDef,
@@ -25,7 +30,10 @@ import {
   locatedError,
 } from 'graphql';
 
-import { GraphQLStreamDirective } from '../type/directives';
+import {
+  GraphQLDeferDirective,
+  GraphQLStreamDirective,
+} from '../type/directives';
 
 import type { Path } from '../jsutils/Path';
 import type { ObjMap } from '../jsutils/ObjMap';
@@ -33,6 +41,7 @@ import type { PromiseOrValue } from '../jsutils/PromiseOrValue';
 import type { Maybe } from '../jsutils/Maybe';
 import type { Push, Stop } from '../jsutils/repeater';
 import { inspect } from '../jsutils/inspect';
+import { memoize1 } from '../jsutils/memoize1';
 import { memoize2 } from '../jsutils/memoize2';
 import { memoize3 } from '../jsutils/memoize3';
 import { invariant } from '../jsutils/invariant';
@@ -56,11 +65,6 @@ import {
   getArgumentValues as _getArgumentValues,
   getDirectiveValues,
 } from './values';
-import type { FieldsAndPatches, PatchFields } from './collectFields';
-import {
-  collectFields,
-  collectSubfields as _collectSubfields,
-} from './collectFields';
 import { mapAsyncIterable } from './mapAsyncIterable';
 import { flattenAsyncIterable } from './flattenAsyncIterable';
 
@@ -120,6 +124,16 @@ interface IncrementalResult {
 interface Publisher {
   push: Push<ExecutionPatchResult>;
   stop: Stop;
+}
+
+export interface PatchFields {
+  label?: string;
+  fields: Map<string, ReadonlyArray<FieldNode>>;
+}
+
+export interface FieldsAndPatches {
+  fields: Map<string, ReadonlyArray<FieldNode>>;
+  patches: Array<PatchFields>;
 }
 
 export interface ExecutorArgs {
@@ -223,17 +237,7 @@ export class Executor {
       exeContext: ExecutionContext,
       returnType: GraphQLObjectType,
       fieldNodes: ReadonlyArray<FieldNode>,
-    ) => {
-      const { fragments, variableValues, enableIncremental } = exeContext;
-      return _collectSubfields(
-        this._executorSchema,
-        fragments,
-        variableValues,
-        returnType,
-        fieldNodes,
-        !enableIncremental,
-      );
-    },
+    ) => this._collectSubfields(exeContext, returnType, fieldNodes),
   );
 
   /**
@@ -256,6 +260,23 @@ export class Executor {
   getFieldDef = memoize2(
     (parentType: GraphQLObjectType, fieldNodes: ReadonlyArray<FieldNode>) =>
       this._getFieldDef(parentType, fieldNodes),
+  );
+
+  /**
+   * Creates a field list, memoizing so that functions operating on the
+   * field list can be memoized.
+   */
+  createFieldList = memoize1((node: FieldNode): Array<FieldNode> => [node]);
+
+  /**
+   * Appends to a field list, memoizing so that functions operating on the
+   * field list can be memoized.
+   */
+  updateFieldList = memoize2(
+    (fieldList: Array<FieldNode>, node: FieldNode): Array<FieldNode> => [
+      ...fieldList,
+      node,
+    ],
   );
 
   private _schema: GraphQLSchema;
@@ -725,13 +746,12 @@ export class Executor {
       );
     }
 
-    const fieldsAndPatches = collectFields(
-      this._executorSchema,
+    const fieldsAndPatches = this.collectFields(
       fragments,
       variableValues,
       rootType,
       operation.selectionSet,
-      !enableIncremental,
+      enableIncremental,
     );
 
     return {
@@ -2187,6 +2207,301 @@ export class Executor {
     }
 
     return value;
+  }
+
+  /**
+   * Given a selectionSet, collects all of the fields and returns them.
+   *
+   * CollectFields requires the "runtime type" of an object. For a field that
+   * returns an Interface or Union type, the "runtime type" will be the actual
+   * object type returned by that field.
+   */
+  collectFields(
+    fragments: ObjMap<FragmentDefinitionNode>,
+    variableValues: { [variable: string]: unknown },
+    runtimeType: GraphQLObjectType,
+    selectionSet: SelectionSetNode,
+    enableIncremental = true,
+  ): FieldsAndPatches {
+    const fields = new Map();
+    const patches: Array<PatchFields> = [];
+    this.collectFieldsImpl(
+      fragments,
+      variableValues,
+      runtimeType,
+      selectionSet,
+      fields,
+      patches,
+      new Set(),
+      enableIncremental,
+    );
+    return { fields, patches };
+  }
+
+  /**
+   * Given an array of field nodes, collects all of the subfields of the passed
+   * in fields, and returns them at the end.
+   *
+   * CollectSubFields requires the "return type" of an object. For a field that
+   * returns an Interface or Union type, the "return type" will be the actual
+   * object type returned by that field.
+   *
+   * @internal
+   */
+  _collectSubfields(
+    exeContext: ExecutionContext,
+    returnType: GraphQLObjectType,
+    fieldNodes: ReadonlyArray<FieldNode>,
+  ): FieldsAndPatches {
+    const subFieldNodes = new Map();
+    const visitedFragmentNames = new Set<string>();
+
+    const subPatches: Array<PatchFields> = [];
+    const subFieldsAndPatches = {
+      fields: subFieldNodes,
+      patches: subPatches,
+    };
+
+    const { fragments, variableValues, enableIncremental } = exeContext;
+
+    for (const node of fieldNodes) {
+      if (node.selectionSet) {
+        this.collectFieldsImpl(
+          fragments,
+          variableValues,
+          returnType,
+          node.selectionSet,
+          subFieldNodes,
+          subPatches,
+          visitedFragmentNames,
+          enableIncremental,
+        );
+      }
+    }
+    return subFieldsAndPatches;
+  }
+
+  collectFieldsImpl(
+    fragments: ObjMap<FragmentDefinitionNode>,
+    variableValues: { [variable: string]: unknown },
+    runtimeType: GraphQLObjectType,
+    selectionSet: SelectionSetNode,
+    fields: Map<string, Array<FieldNode>>,
+    patches: Array<PatchFields>,
+    visitedFragmentNames: Set<string>,
+    enableIncremental: boolean,
+  ): void {
+    for (const selection of selectionSet.selections) {
+      switch (selection.kind) {
+        case Kind.FIELD: {
+          if (!this.shouldIncludeNode(variableValues, selection)) {
+            continue;
+          }
+          const name = this.getFieldEntryKey(selection);
+          const fieldList = fields.get(name);
+          if (fieldList !== undefined) {
+            fields.set(name, this.updateFieldList(fieldList, selection));
+          } else {
+            fields.set(name, this.createFieldList(selection));
+          }
+          break;
+        }
+        case Kind.INLINE_FRAGMENT: {
+          if (
+            !this.shouldIncludeNode(variableValues, selection) ||
+            !this.doesFragmentConditionMatch(selection, runtimeType)
+          ) {
+            continue;
+          }
+
+          const defer = this.getDeferValues(
+            variableValues,
+            selection,
+            enableIncremental,
+          );
+
+          if (defer) {
+            const patchFields = new Map();
+            this.collectFieldsImpl(
+              fragments,
+              variableValues,
+              runtimeType,
+              selection.selectionSet,
+              patchFields,
+              patches,
+              visitedFragmentNames,
+              enableIncremental,
+            );
+            patches.push({
+              label: defer.label,
+              fields: patchFields,
+            });
+          } else {
+            this.collectFieldsImpl(
+              fragments,
+              variableValues,
+              runtimeType,
+              selection.selectionSet,
+              fields,
+              patches,
+              visitedFragmentNames,
+              enableIncremental,
+            );
+          }
+          break;
+        }
+        case Kind.FRAGMENT_SPREAD: {
+          const fragName = selection.name.value;
+
+          if (!this.shouldIncludeNode(variableValues, selection)) {
+            continue;
+          }
+
+          const defer = this.getDeferValues(
+            variableValues,
+            selection,
+            enableIncremental,
+          );
+          if (visitedFragmentNames.has(fragName) && !defer) {
+            continue;
+          }
+
+          const fragment = fragments[fragName];
+          if (
+            !fragment ||
+            !this.doesFragmentConditionMatch(fragment, runtimeType)
+          ) {
+            continue;
+          }
+          visitedFragmentNames.add(fragName);
+
+          if (defer) {
+            const patchFields = new Map();
+            this.collectFieldsImpl(
+              fragments,
+              variableValues,
+              runtimeType,
+              fragment.selectionSet,
+              patchFields,
+              patches,
+              visitedFragmentNames,
+              enableIncremental,
+            );
+            patches.push({
+              label: defer.label,
+              fields: patchFields,
+            });
+          } else {
+            this.collectFieldsImpl(
+              fragments,
+              variableValues,
+              runtimeType,
+              fragment.selectionSet,
+              fields,
+              patches,
+              visitedFragmentNames,
+              enableIncremental,
+            );
+          }
+          break;
+        }
+      }
+    }
+  }
+
+  /**
+   * Returns an object containing the `@defer` arguments if a field should be
+   * deferred based on the experimental flag, defer directive present and
+   * not disabled by the "if" argument.
+   */
+  getDeferValues(
+    variableValues: { [variable: string]: unknown },
+    node: FragmentSpreadNode | InlineFragmentNode,
+    enableIncremental: boolean,
+  ): undefined | { label?: string } {
+    if (!enableIncremental) {
+      return;
+    }
+
+    const defer = getDirectiveValues(
+      this._executorSchema,
+      GraphQLDeferDirective,
+      node,
+      variableValues,
+    );
+
+    if (!defer) {
+      return;
+    }
+
+    if (defer.if === false) {
+      return;
+    }
+
+    return {
+      label: typeof defer.label === 'string' ? defer.label : undefined,
+    };
+  }
+
+  /**
+   * Determines if a field should be included based on the `@include` and `@skip`
+   * directives, where `@skip` has higher precedence than `@include`.
+   */
+  shouldIncludeNode(
+    variableValues: { [variable: string]: unknown },
+    node: FragmentSpreadNode | FieldNode | InlineFragmentNode,
+  ): boolean {
+    const skip = getDirectiveValues(
+      this._executorSchema,
+      GraphQLSkipDirective,
+      node,
+      variableValues,
+    );
+    if (skip?.if === true) {
+      return false;
+    }
+
+    const include = getDirectiveValues(
+      this._executorSchema,
+      GraphQLIncludeDirective,
+      node,
+      variableValues,
+    );
+    if (include?.if === false) {
+      return false;
+    }
+    return true;
+  }
+
+  /**
+   * Determines if a fragment is applicable to the given type.
+   */
+  doesFragmentConditionMatch(
+    fragment: FragmentDefinitionNode | InlineFragmentNode,
+    type: GraphQLObjectType,
+  ): boolean {
+    const typeConditionNode = fragment.typeCondition;
+    if (!typeConditionNode) {
+      return true;
+    }
+    const conditionalType = this._executorSchema.getType(typeConditionNode);
+    if (conditionalType === type) {
+      return true;
+    }
+    if (
+      conditionalType &&
+      this._executorSchema.isAbstractType(conditionalType)
+    ) {
+      return this._executorSchema.isSubType(conditionalType, type);
+    }
+    return false;
+  }
+
+  /**
+   * Implements the logic to compute the key of a given field's entry
+   */
+  getFieldEntryKey(node: FieldNode): string {
+    return node.alias ? node.alias.value : node.name.value;
   }
 }
